+++ src/plugins/04b03.ts
@@ -3,6 +3,10 @@
 import KaboomCtx from "../ctx";
 import b04b03Src from "./04b03_6x8.png";
 
+import {
+	FontData,
+} from "../assets";
+
 module.exports = (k: KaboomCtx) => {
 
 function load04b03(): Promise<FontData> {